Add core runtime controller tests (#269)

+++ pkg/controllers/core/deployer_controller.go
@@ -90,7 +90,7 @@ func (r *DeployerReconciler) Reconcile(req ctrl.Request) (ctrl.Result, error) {
 	result, err := r.reconcile(ctx, log, &deployer)
 
 	// check if status has changed before updating, unless requeued
-	if !result.Requeue && !equality.Semantic.DeepEqual(deployer.Status, originalDeployer.Status) {
+	if !result.Requeue && !equality.Semantic.DeepEqual(deployer.Status, originalDeployer.Status) && deployer.GetDeletionTimestamp() == nil {
 		// update status
 		log.Info("updating deployer status", "diff", cmp.Diff(originalDeployer.Status, deployer.Status))
 		if updateErr := r.Status().Update(ctx, &deployer); updateErr != nil {
@@ -410,10 +410,12 @@ func (r *DeployerReconciler) reconcileIngress(ctx context.Context, log logr.Logg
 
 	// delete ingress if no longer needed
 	if desiredIngress == nil {
-		log.Info("deleting ingress", "ingress", actualIngress)
-		if err := r.Delete(ctx, &actualIngress); err != nil {
-			log.Error(err, "unable to delete ingress for Deployer", "ingress", actualIngress)
-			return nil, err
+		if !actualIngress.CreationTimestamp.IsZero() {
+			log.Info("deleting ingress", "ingress", actualIngress)
+			if err := r.Delete(ctx, &actualIngress); err != nil {
+				log.Error(err, "unable to delete ingress for Deployer", "ingress", actualIngress)
+				return nil, err
+			}
 		}
 		return nil, nil
 	}
